Add standard builtin functions, and basic code cleanup
This change touches pretty much everything, but the major change is the
addition of standard builtin hasp functions in Builtins.hs, that are
easiest to implement in Haskell. There will be more builtin functions to
come, however most of them will rely on syntactic forms such as if
statements, lambda and define, as well as the ability to include (by
default) standard library functions that are written in hasp.