Use Data.Sequence instead of List in tokenizer and parser
The tokenizer and parser generate sequences of tokens and expressions by
appending to the end of an accumulator, which in Haskell is inefficient
when the accumulator is of type list, doing O(n) operations every time
an element is appended to the end of the list. Instead we use
Data.Sequence, which implements the same operations in O(1) amortized
time, and convert to a list when done, which is a one-time cost of O(n).